refactor: split cmd root and keep domain ids raw
- split cmd root logic into flags/run/io/json files
- remove tab/url formatting concerns from internal/niconico
- keep GetVideoList and sort working on raw sm* IDs
- scope gofmt checks to tracked Go files in CI/release/docs
- update design documentation for the new responsibility boundary

AI-Assisted: Codex
