+++ .github/workflows/ci.yml
@@ -33,7 +33,7 @@ jobs:
 
       - name: Check gofmt
         run: |
-          gofmt -w .
+          gofmt -w $(git ls-files '*.go')
           git diff --exit-code
 
       - name: Go vet
