+++ .github/workflows/release.yml
@@ -28,7 +28,7 @@ jobs:
           git diff --exit-code
       - name: Check gofmt
         run: |
-          gofmt -w .
+          gofmt -w $(git ls-files '*.go')
           git diff --exit-code
       - name: Go vet
         run: go vet ./...
